refactor: relocate server probing logic to a dedicated internal/probe package
