+++ internal/download/manager.go
@@ -14,7 +14,7 @@ import (
 	"github.com/SurgeDM/Surge/internal/engine/events"
 	"github.com/SurgeDM/Surge/internal/engine/single"
 	"github.com/SurgeDM/Surge/internal/engine/types"
-	"github.com/SurgeDM/Surge/internal/processing"
+	"github.com/SurgeDM/Surge/internal/probe"
 	"github.com/SurgeDM/Surge/internal/utils"
 )
 
@@ -71,8 +71,8 @@ func uniqueFilePath(path string) string {
 	return path
 }
 
-// TUIDownload is the main entry point for downloads executed by the Engine pool
-func TUIDownload(ctx context.Context, cfg *types.DownloadConfig) error {
+// RunDownload is the main entry point for downloads executed by the Engine pool
+func RunDownload(ctx context.Context, cfg *types.DownloadConfig) error {
 	start := time.Now()
 	if cfg.Runtime == nil {
 		cfg.Runtime = types.DefaultRuntimeConfig()
@@ -181,7 +181,7 @@ func TUIDownload(ctx context.Context, cfg *types.DownloadConfig) error {
 				ProxyURL:  cfg.Runtime.ProxyURL,
 				CustomDNS: cfg.Runtime.CustomDNS,
 			}
-			valid, errs := processing.ProbeMirrorsWithProxy(ctx, allToCheck, runCfg)
+			valid, errs := probe.ProbeMirrorsWithProxy(ctx, allToCheck, runCfg)
 
 			// Log errors
 			for u, e := range errs {
@@ -307,5 +307,5 @@ func Download(ctx context.Context, url string, outPath string, progressCh chan<-
 	}
 	// Default runtime config
 	cfg.Runtime = types.DefaultRuntimeConfig()
-	return TUIDownload(ctx, &cfg)
+	return RunDownload(ctx, &cfg)
 }
