+++ internal/download/mirror_resume_test.go
@@ -106,7 +106,7 @@ func TestIntegration_MirrorResume(t *testing.T) {
 	// Start download and interrupt
 	errCh := make(chan error)
 	go func() {
-		errCh <- download.TUIDownload(ctx1, &cfg)
+		errCh <- download.RunDownload(ctx1, &cfg)
 	}()
 
 	// Wait until download really started so Pause() has an attached cancel func.
@@ -188,10 +188,10 @@ func TestIntegration_MirrorResume(t *testing.T) {
 		Mirrors:       []string{}, // Empty mirrors!
 	}
 
-	// We can't easily hook into TUIDownload to verify it loaded mirrors without running it.
+	// We can't easily hook into RunDownload to verify it loaded mirrors without running it.
 	ctx2 := context.Background()
 	go func() {
-		errCh <- download.TUIDownload(ctx2, &resumeCfg)
+		errCh <- download.RunDownload(ctx2, &resumeCfg)
 	}()
 
 	// Give it enough time to start and restore mirrors from saved state.
