+++ internal/download/resume_test.go
@@ -96,7 +96,7 @@ func TestIntegration_PauseResume(t *testing.T) {
 	// Start download
 	errCh := make(chan error)
 	go func() {
-		errCh <- download.TUIDownload(ctx, &cfg)
+		errCh <- download.RunDownload(ctx, &cfg)
 	}()
 
 	// Wait for some progress
@@ -176,7 +176,7 @@ func TestIntegration_PauseResume(t *testing.T) {
 	// Reset pause flag before resume
 	progState.Resume()
 
-	err = download.TUIDownload(resumeCtx, &cfg)
+	err = download.RunDownload(resumeCtx, &cfg)
 	if err != nil {
 		t.Fatalf("Resume failed: %v", err)
 	}
