+++ internal/probe/probe.go
@@ -1,4 +1,4 @@
-package processing
+package probe
 
 import (
 	"context"
